requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
plotly==5.22.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tenacity==8.3.0
tzdata==2024.1
Werkzeug==3.0.3
